Deal with multiple entries of same task in engine list.
Tasks are added to engine every time current_engine is
set, for example when running cont() after an idle();
and since cont() also calls multiplex() it can easily
happen that a task is added to engines that shouldn't
run it, or is added more than once to the same task.

This patch causes multiplex() to return when being
called from an engine after current_engine was already
changed again.