Remove cont() and lots of other WIP
After this, we only have one control function
left: signalled(), and even that shouldn't be
called directly; instead call condition.signal();

idle() and cont() were removed. The go idle,
you now have to call wait(condition), where
condition is an AICondition (which was rewritten
from scratch and is kinda a different beast now,
without queueing).

An AICondition now has to be constructed with as
parameter a pointer to a StatefulTask, which
then uses it in a single state to pass it to wait().
After that the condition should not be used again.