Get rid of AICondition.
Because the AICondition object was referenced by both,
the code signalling it as well as the task waiting for it,
we had a serious life-time problem: The AICondition object
kept a boost::intrusive_ptr<AIStatefulTask> to the task
it was signalling upon a call to signal(), while it was
owned by that very same Task (usually), causing such a
task to be never freed.

The only way around this would be to add reference counting
to the AICondition objects too and then have both, the
code that signals an AICondition as well as the Task
waiting for it to explicitly reset their reference once they
are done with it.

All this complexity now has been avoided by changing a
condition object in a single bit; this limits the number
of conditions that one task can wait for to 32 (or 64 if
we change idle_mask_type to a uint64_t) but that isn't
much of a downside.

Each Task now keeps three 32-bit masks as part of their
substate: not_idle, skip_idle and idle where each bit
represents a 'Condition' object. The code previously
signalling the Condition object now signals the (waiting)
Task directly, which also requires boost::intrusive_ptr,
but just for the AIStatefulTask, which is by not as intrusive
as a new reference counted object ;).