Make non-key part of std::set mutable.
If you use a std::set as a 'map' (only part of the
elements of the set are used to order the set) then
you run into the problem that dereferencing the
set iterator always gives you a const element (because
if you'd change the type that the iterator is pointing
to you could change how the element should be ordered
inside the set...) But because IS ok to change those
member variables that are NOT part of the sorting
function (operator<) the usual trick is to make those
'mutable', so you can change them even though the object
is const. Of course this has a downside too: you lose
the protection that the 'const' keyword gives you normally.