Bug fix.
It was not allowed to call finish() when the task is idle (wait()
was called). But that seems completely legal when only
wait(slow_down_condition) was called because this task is the parent
of a task that failed to add something to the thread pool.
We can savely finish in that case: there is no threat that we'll
be flooding the thread pool with more.